malaria_ast_starter/malaria_ast_starter/requirements.txt
torch
torchvision
tqdm
numpy
pillow
pyyaml
scikit-learn
matplotlib
seaborn
adaptive-sparse-training
